Add position for chapters for consistency with sections
Also treat position as one based index, there is no need to be zero
based, if that would be the case use any index instead.

The story elements position is abstracted in a way that the value is
owned by a persistent field, instead of relying directly in the
relationship with the index in every place it is being implemented.