Angular rewrite of story viewer
The story viewer using jquery ui was unmaintainable, impossible to
create anything else. It took more than 1 month to rewrite this, but I
believe now this is a good base to be evolved on.

impress.js was changed to jmpress, using a customized fork and an
angular port.
Now it seems to be working in opera.
It wasn't tested in a real mobile, but probably works well (if it
doesn't, just needs a few adjustments)

Closes gh-49. Closes gh-45.