Replace the FB Like Box Plugin to the FB Page Plugin
When The Graph API v2.3 was released, the Facebook Like Box plugin was
deprecated. In June 23rd 2015 all Like Box implementations started using
the Page Plugin automatically.

This commit properly adjusts the Page Plugin with the new configs for
all screen sizes instead of relying in the fall back from the Like Box.